Fix overflow wrap
Pagefind is hard to style directly. It can inherit from your styles, but that looked a bit off in our case; otherwise by default it has a "style reset" so its styles are encapsulated and not affected by other styles. That means we can't style something like .pagefind-ui__result directly, and because the nodes don't exist right away (they are added dynamically) we can't target them at DOM load time with Javascript.

We use a MutationObserver to see when nodes are added to the DOM and target them with styles at this point.